+++ lydian/const.py
@@ -86,6 +86,8 @@ EMBED_COLOR_ERROR: int = 0xff0000
 
 # Compiled regex
 COLOR_ESCAPE_REGEX: re.Pattern[str] = re.compile(r'\x1b\[.*?m')
+HTTP_REGEX: re.Pattern[str] = re.compile(r'^https?://')
+"""Matches if a string begins with ``http://`` or ``https://``."""
 DOCSTRING_PARAM_REGEX: re.Pattern[str] = re.compile(
     r'^:param (?P<name>\w+): (?P<desc>.+(?:\n    .+|\n)*)',
     flags=re.MULTILINE,
@@ -97,6 +99,14 @@ Named groups:
     - "desc"
 """
 YTDL_DOWNLOAD_PROGRESS_REGEX: re.Pattern[str] = re.compile(r'\[download\].+ETA')
+YTDL_SEARCH_PREFIX_REGEX: re.Pattern[str] = re.compile(r'(?P<prefix>.*search)(?P<n>\d*):(?P<query>.*)')
+"""Matches yt-dlp search extractor prefixes like "ytsearch" and "scsearch".
+
+Named groups:
+    - "prefix": The search prefix, excluding the following number
+    - "n": (optional) The number of results to get, if present
+    - "query": The string following the prefix and colon
+"""
 
 MD_HEADER_REGEX: re.Pattern[str] = re.compile(r'^#(?P<title>.*)$', flags=re.MULTILINE)
 """Matches a markdown header with any number of ``#`` characters.
@@ -127,6 +137,19 @@ Named groups:
 USER_AGENT: str = f'lydian-discord-bot/{__version__}'
 DEFAULT_DISCORD_PROMPT_TIMEOUT: float = 60.0
 DEFAULT_DISCORD_PAGINATED_VIEW_TIMEOUT: float = 60.0 * 5
+EMOJI_DIGITS: tuple[str, str, str, str, str, str, str, str, str, str] = (
+    emojize(':zero:', language='alias'),
+    emojize(':one:', language='alias'),
+    emojize(':two:', language='alias'),
+    emojize(':three:', language='alias'),
+    emojize(':four:', language='alias'),
+    emojize(':five:', language='alias'),
+    emojize(':six:', language='alias'),
+    emojize(':seven:', language='alias'),
+    emojize(':eight:', language='alias'),
+    emojize(':nine:', language='alias'),
+)
+"""Digits 0-9 as emoji."""
 QUEUE_MAX_PER_PAGE: int = 20
 
 class ConsoleHighlighter(Highlighter):
@@ -159,6 +182,12 @@ class EmojiStr(StrEnum):
     SHUFFLE = emojize(':twisted_rightwards_arrows:', language='alias')
     LOOP     = emojize(':repeat:', language='alias')
     LOOP_ONE = emojize(':repeat_one:', language='alias')
+
+    # Alphanumeric
+    @classmethod
+    def from_int(cls, n: int) -> str:
+        """Converts an integer into digit emojis."""
+        return ''.join(EMOJI_DIGITS[int(ch)] for ch in str(n).lower())
 
 class LogLevel(IntEnum):  # noqa: D101
     TRACE   = 5